+++ app/api/events.rb
@@ -1,5 +1,24 @@
+module ParamsHelper
+  extend Grape::API::Helpers
+
+  params :filters do
+    optional :all,
+             type: Boolean,
+             desc: 'весь список, включая завершенные'
+  end
+end
+
+module FiltersHelper
+  extend Grape::API::Helpers
+
+  def events_scope(all)
+    scope = Event.order(:id)
+    all ? scope : scope.where(done: false)
+  end
+end
+
 class Events < Grape::API
-  helpers FiltersHelper, EventsHelper
+  helpers FiltersHelper, ParamsHelper
 
   resource :events do
     desc 'Список дел'
