--- app/api/helpers/events_helper.rb
@@ -1,8 +1,0 @@
-module EventsHelper
-  extend Grape::API::Helpers
-
-  def events_scope(all)
-    scope = Event.order(:id)
-    all ? scope : scope.where(done: false)
-  end
-end
