+++ app/api/helpers/params_helper.rb
@@ -1,9 +1,9 @@
-module ParamsHelper
-  extend Grape::API::Helpers
+# module ParamsHelper
+#   extend Grape::API::Helpers
 
-  params :filters do
-    optional :all,
-             type: Boolean,
-             desc: 'весь список, включая завершенные'
-  end
-end
+#   params :filters do
+#     optional :all,
+#              type: Boolean,
+#              desc: 'весь список, включая завершенные'
+#   end
+# end
