+++ app/api/root_api.rb
@@ -5,7 +5,7 @@ class RootApi < Grape::API
   format :json
   prefix :api
 
-  #mount Events
+  mount Events
 
   #mount Events
